+++ src/nu/validator/datatype/SandboxAllowList.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Mozilla Foundation
+ * Copyright (c) 2017 Mozilla Foundation
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -47,6 +47,10 @@ public final class SandboxAllowList extends AbstractDatatype {
 
     private boolean hasAllowSameOrigin;
 
+    private boolean hasAllowTopNavigation;
+
+    private boolean hasAllowTopNavigationByUserActivation;
+
     static {
         allowedKeywords.add("allow-forms");
         allowedKeywords.add("allow-modals");
@@ -58,6 +62,7 @@ public final class SandboxAllowList extends AbstractDatatype {
         allowedKeywords.add("allow-same-origin");
         allowedKeywords.add("allow-scripts");
         allowedKeywords.add("allow-top-navigation");
+        allowedKeywords.add("allow-top-navigation-by-user-activation");
     }
 
     @Override
@@ -87,6 +92,15 @@ public final class SandboxAllowList extends AbstractDatatype {
                             + " embedded page to break out of all sandboxing.",
                     WARN);
         }
+        if (hasAllowTopNavigation //
+                && hasAllowTopNavigationByUserActivation) {
+            throw newDatatypeException(
+                    "\u201callow-top-navigation-by-user-activation\u201d and"
+                            + " \u201callow-top-navigation\u201d must not both"
+                            + " be specified at the same time. If both are"
+                            + " present, only \u201callow-top-navigation\u201d"
+                            + " will have effect.");
+        }
     }
 
     private void checkToken(CharSequence literal, StringBuilder builder, int i,
@@ -106,6 +120,12 @@ public final class SandboxAllowList extends AbstractDatatype {
         }
         if ("allow-same-origin".equals(token)) {
             hasAllowSameOrigin = true;
+        }
+        if ("allow-top-navigation".equals(token)) {
+            hasAllowTopNavigation = true;
+        }
+        if ("allow-top-navigation-by-user-activation".equals(token)) {
+            hasAllowTopNavigationByUserActivation = true;
         }
     }