+++ CHANGELOG.md
@@ -5,6 +5,7 @@ changes to any options/interfaces the checker exposes for developers.
 # 17.N.N
 NN March 2017
   - Allow `meta[name]` to have `itemref`/`itemscope`/`itemtype`/`itemid`
+  - Allow `allow-top-navigation-by-user-activation` in `iframe[sandbox]`
   - Stop hiding “sectioning roots” headings in “Heading-level outline”
   - Make “Corrupt GZIP trailer” a non-error
   - Add `--asciiquotes` option to vnu.jar command-line checker